Fix/Improve Makefile + add debug/visibility
We can now build libft.a with -fvisibility=hidden when creating a new
library on top of it in order to hide symbols from libft in the newly
created library.
To use this flag just add `LIBFT_HIDDEN=TRUE` to the make cmdline.

We can now also build libft.a in debug mode - `-O0 -g` - to ease debugging
the libft or another project which use libft.
To use this those flags just add `DEBUG=TRUE` to the make cmdline.